now using TSV writer to create variant file, combined HotspotVariantComposition and VariantComposition classes
